Extract shared league team-id helpers
