+++ src/commandsHandler/leagueGraphHandler.js
@@ -7,6 +7,7 @@ const { getLeagueData } = require('../azureStorageService');
 const { fetchCurrentSeasonRaces } = require('../raceScheduleService');
 const { getChipEmoji } = require('../utils/chipEmojis');
 const { getSelectedTeam } = require('../cache');
+const { buildTeamId } = require('../utils/teamId');
 const {
   LEAGUE_GRAPH_CALLBACK_TYPE,
   COMMAND_FOLLOW_LEAGUE,
@@ -27,22 +28,6 @@ const TEAM_COLOR_PALETTE = [
   '#000075', // navy
   '#a9a9a9', // grey
 ];
-
-function sanitizeTeamName(name) {
-  const base = String(name || 'team')
-    .normalize('NFKD')
-    .replace(/[^\w-]+/g, '-')
-    .replace(/-+/g, '-')
-    .replace(/^-|-$/g, '');
-
-  const trimmed = base.length > 0 ? base : 'team';
-
-  return trimmed.slice(0, 40);
-}
-
-function buildTeamId(leagueCode, teamName) {
-  return `${leagueCode}_${sanitizeTeamName(teamName)}`;
-}
 
 /**
  * Extract matchday keys from a teams array and sort them by trailing numeric id.
